feat: Add stage labels to ring charts as well
We'll ask what they prefer
