+++ src/components/StagesDescription.vue
@@ -12,7 +12,7 @@
           :src="getStepLogo(step)"
           :alt="step.name + ' illustration'"
         />
-        <div class="font-semibold">{{ PRETTY_NAME_BY_STEP[step.name] }}</div>
+        <div class="font-semibold">{{ getStageLabel(step.name) }}</div>
         <p class="text-center">{{ step.description || 'Pas de description' }}</p>
       </div>
     </div>
@@ -29,25 +29,15 @@ import WholesaleLogo from '../images/icons/wholesale.svg'
 import CollectionLogo from '../images/icons/collection.svg'
 import ProductionLogo from '../images/icons/production.svg'
 import ExportLogo from '../images/icons/export.svg'
+import { STAGES as DISPLAY_STAGES, getStageLabel } from '@utils/stages'
 
 const props = defineProps({
     stagesDescription: Object
 })
 
-const PRETTY_NAME_BY_STEP = {
-  'Producers': 'Production',
-  'Collectors': 'Collection',
-  'Processors': 'Processing',
-  'Wholesalers': 'Wholesaling',
-  'Retailers': 'Retailing',
-  'End use': 'End use'
-};
-
-const DISPLAY_STEPS = Object.keys(PRETTY_NAME_BY_STEP);
-
 const populatedSteps = computed(() => {
   return props.stagesDescription
-    .filter((stage) => DISPLAY_STEPS.includes(stage.name))
+    .filter((stage) => DISPLAY_STAGES.includes(stage.name))
     .map((stage) => ({
       ...stage,
       image: ``
